+++ src/test/groovy/com/palantir/gradle/externalpublish/ExternalPublishGradlePluginPluginSpec.groovy
@@ -17,15 +17,19 @@
 package com.palantir.gradle.externalpublish
 
 import com.fasterxml.jackson.core.type.TypeReference
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.palantir.gradle.publish.GradlePluginDef
 import nebula.test.IntegrationSpec
 import nebula.test.functional.ExecutionResult
 
 class ExternalPublishGradlePluginPluginSpec extends IntegrationSpec {
+    static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+
     File buildDir
     File pomFile
     File metaFilesPath
 
-    static final TypeReference<List<GenerateGradlePluginMetaDataTask.GradlePluginDef>> GRADLE_PLUGIN_DEF_TYPE_REF = new TypeReference<List<GenerateGradlePluginMetaDataTask.GradlePluginDef>>() {};
+    static final TypeReference<List<GradlePluginDef>> GRADLE_PLUGIN_DEF_TYPE_REF = new TypeReference<List<GradlePluginDef>>() {};
 
     def setup() {
         buildDir = new File(projectDir, "build")
@@ -81,7 +85,7 @@ class ExternalPublishGradlePluginPluginSpec extends IntegrationSpec {
 
         def xml = new XmlSlurper().parseText(pomFile.text)
         def propString = xml.properties[ExternalPublishGradlePluginPlugin.PUBLISHED_PLUGIN_IDS_KEY].toString()
-        List<GenerateGradlePluginMetaDataTask.GradlePluginDef> publishedPlugins = GenerateGradlePluginMetaDataTask.OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
+        List<GradlePluginDef> publishedPlugins = OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
 
         publishedPlugins[0].id == 'com.palantir.test-plugin1'
         publishedPlugins[0].implementingClass == 'com.palantir.gradle.TestPlugin1'
@@ -105,7 +109,7 @@ class ExternalPublishGradlePluginPluginSpec extends IntegrationSpec {
 
         def xml = new XmlSlurper().parseText(pomFile.text)
         def propString = xml.properties[ExternalPublishGradlePluginPlugin.PUBLISHED_PLUGIN_IDS_KEY].toString()
-        List<GenerateGradlePluginMetaDataTask.GradlePluginDef> publishedPlugins = GenerateGradlePluginMetaDataTask.OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
+        List<GradlePluginDef> publishedPlugins = OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
 
         publishedPlugins[0].id == 'com.palantir.test-plugin1'
         publishedPlugins[0].implementingClass == 'com.palantir.gradle.TestPlugin1'
@@ -144,7 +148,7 @@ class ExternalPublishGradlePluginPluginSpec extends IntegrationSpec {
         // Both properties are present in the same block
         propertiesBlock.'existing-property'.text() == 'existing-value'
         def propString = propertiesBlock[ExternalPublishGradlePluginPlugin.PUBLISHED_PLUGIN_IDS_KEY].text()
-        List<GenerateGradlePluginMetaDataTask.GradlePluginDef> publishedPlugins = GenerateGradlePluginMetaDataTask.OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
+        List<GradlePluginDef> publishedPlugins = OBJECT_MAPPER.readValue(propString, GRADLE_PLUGIN_DEF_TYPE_REF)
         publishedPlugins[0].id == 'com.palantir.test-plugin1'
         publishedPlugins[0].implementingClass == 'com.palantir.gradle.TestPlugin1'
     }
