+++ versions.lock
@@ -69,7 +69,7 @@ org.apiguardian:apiguardian-api:1.1.2 (6 constraints: 5366ce6e)
 org.codehaus.groovy:groovy:3.0.25 (3 constraints: 14347add)
 org.hamcrest:hamcrest:2.2 (1 constraints: d20cdc04)
 org.hamcrest:hamcrest-core:1.3 (1 constraints: cc05fe3f)
-org.junit.jupiter:junit-jupiter:5.13.4 (1 constraints: 11051e36)
+org.junit.jupiter:junit-jupiter:5.13.4 (1 constraints: 7e074077)
 org.junit.jupiter:junit-jupiter-api:5.13.4 (4 constraints: be46df8d)
 org.junit.jupiter:junit-jupiter-engine:5.13.4 (1 constraints: 3a0e0a4a)
 org.junit.jupiter:junit-jupiter-params:5.13.4 (2 constraints: f82423fc)
